+++ src/Bespoken.php
@@ -34,7 +34,7 @@ use yii\log\Logger;
  */
 class Bespoken extends Plugin
 {
-    public string $schemaVersion = '1.0.0';
+    public string $schemaVersion = '1.1.0';
     public bool $hasCpSettings = true;
     public bool $hasCpSection = false;